refactor: Changed CWS references to WebApi, re-organized Web folder
